The flatMap operator
Where the map operator takes each emission and transforms them, the flatMap operator
creates a new producer, applying the function you passed to each emission of the source
producer.

The output is similar to the previous one, but the logic is different. Instead of just returning
the String, we are returning Observable with the desired String. Although, for this
example, you seem to have no benefit using it, think of a situation when you need to derive
multiple items from a single emission.