The zipWith operator
The instance version (that is, the copy of the function, which should be called with an
instance rather than static) of the zip operator is zipWith, which can be called from the
Observable instance itself. The only problem with this version is that you can pass only
another source Observable. If you need to work with three or more Observable
instances, you should rather consider using the zip operator instead of zipWith.