The combineLatest operator
The combineLatest operator works in a similar way like the zip operator. It accumulates
the emissions of the provided producers. The only difference between combineLatest and
zip is that the zip operator waits for each of its source producers to emit, before it starts
processing all the emissions to create its new one, but the combineLatest operator starts as
soon as it receives any emit from any of its source producers.