Testing RxKotlin Applications
This chapter is dedicated to testing. As Kotlin itself is relatively new, our first objective
would be to learn testing in Kotlin. We will then proceed with testing in RxKotlin. The
following are the topics we are going to cover in this chapter:
Introduction to unit testing and its importance
Kotlin and JUnit, Kotlin-test
Testing tools in RxKotlin
Blocking subscribers
Blocking operators
TestObserver and TestSubscriber